gmgrl - starting to read instructions
Starting to read CPU instructions. Here we are initializing the cpu and
registers after the boot load sequence. We are implementing the
instructions one at a time and can look at abstractions later once the
patterns become more obvious. The instructions will often have common
logic utilizing different registers, or value sizes, but we can keep it
simple to begin with. Included in this change set is the ability to step
through the program one instruction at a time which has been useful for
debugging.