main.c - add the ability to continue to the next stop
Add the ability to run using 'c' until the next stop (in our current
case that indicates an unimplemented opcode). This lets us run up until
a specific point, without needing to step each instruction. Once a stop
happens we can contiue stepping as needed, or resume again using the
continue key. In the interim this also provides an easy mechansim to
hack in breakpoints, but in the long term it should also facilitate
breakpoint support by running in a very similar loop.