viewer - added screen and simple viewer program
Adding a screen struct, and viewer program. The screen currently just
draws the gameboy pixels into an file using the fg and bg ansi sequences
to draw the top and bottom halves of the character in two different
colors. This effectively gives us a pixel grid that we can view in the
terminal. The viewer program is a simple program which outputs the file
continously into a terminal window. This gives us a quick and dirty way
of viewing the screen without needing to write an entire rendered gui.