viewer - remove the viewer and move towards ppu instead
The goal with the viewer was for us to provide a lightweight display
outside of the main engine. It turned out to be a failed experiment as
we really implemented it ahead of the ppu. Instead let's move towards a
simplified approach just using the same console. By the time we need
more we should invest in a gui anyway.