gmgrl - ppu beginings
This is only tested in that we don't crash. We likely already have
erreonous logic, but it's hard for us to track that down at the moment.
This get's us some scaffolding at the very least and we can begin to
tease out the finer details from here. We're hitting a limit to our
ability to just printf debug errors, and so we are likely due for some
form of test infrastructure.